% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRemover.R
\name{get_diff}
\alias{get_diff}
\title{Calculates the average load difference between the cell-cycle genes and
control genes for each component.}
\usage{
get_diff(xy, xn)
}
\arguments{
\item{xy}{The data for the genes which are annotated to the cell-cycle, i.e.
those genes for which "if_cc" is \code{TRUE}.}

\item{xn}{The data for the genes which are not annotated to the cell-cycle,
control genes, genes for which "if_cc" is \code{FALSE}}
}
\value{
A data frame containing the loadings for each component on the
cell-cycle and control genes.
}
\description{
This is an interal function for use by "bootstrap_diff" only.
}
