\name{t.cell_data}
\alias{t.cell_data}
\docType{data}
\title{ Single-Cell RNA-Seq data from differentiating T-helper cells}
\description{
  This data contains expression levels (log-transformed normalized count values)
  for 81 cells and 14,147 genes. The data was normalized using ERCC spike-ins. This
  data was generated by Mahata, B. et al (2014). The processed data was retrieved from
  the supplementary material of Buettner et al. (2015), for descriptions of how the data was
  processed, please refer to their paper.
}
\usage{data(t.cell_data)}
\format{
  A data set that contains with the following attributes
  \describe{
    \item{\code{t.cell_data}}{the data matrix. rows are cells, and columns are genes. These should be treated as log-transformed and normalized}
  }
}
\value{A scRNA-Seq dataset with gene expression levels for 187 T-helper cells}
\keyword{datasets}
